Introduced core.log to the project.
In iFlux the log functions used to be in conf. Here we're placing
them in core.log. The issue before was the way conf used to be
handled, loading a lot of 3th parties classes and functions that
used to parse custom configuration.

This issue was fixed with yaml so we can place the log
functionalities in the right place.

Fixes: #16