Finished the application run task. Loaded enabled components and handlers during the application bootstrap.
The applicatio run task is running correctly and handlers
registered on the library and application configuration are
resolving too.

To get this working the enabled components on conf.components were
loaded during the application bootstrap and the handlers registere
on them were added to the Tornado application.

Used the pythonpath insted of python_path on the application run
task.

Refers: #15
Fixes: #14