Introduced connectors to the data module.
The connectors used to be called connection handlers. On Firenado
the terminology will change to make more sense and don't make
confusion between entities.

The term "handlers" will be used only on the http handlers, from
the tornado web server for instance.

This commit is a basic migration from iFlux to Firenado and despite
of something must be changed, a better review should be made on
the data.connector module.

Only the redis connector was tested.

Refers: #20.