Closed the test sqlalchemy connection instead of session.
When trying to close session a lot of errors were being logged on
the console. Now the connection is being closed so the resource go
back to the engine that will be disposed and close all connections
on the pool.

This issue needs more attention when we start to get more pooling
details implemented.

Refs: #127