Implemented pessimistic connection handler from sqlalchemy doc.
The solution will cover the session errors after a time with the
app online and will make the connection pool work properly.

Now the get connection will not hold the engine and session and
the session property will return a new session. It is necessary
to close the session after using it or else the connection pool
will be used entirely and your app will not work anymore.

Fixes: #145