Redis is expiring session right now.
We don't fisically remove the session thus we declare them with a
expire time on redis. Redis will clear them for us.

What we do is check of sessions without ttls and set a expire time.
This is an overkill, so we can set the session scan interval pretty
high in this case.

Refs: #196