Implemented file session expiration.
Session data and the last write is being stored into the file
separated by '--'. The serialized data was converted to hexadecimal
to ensure that no '--' would apear in the data part.

The file session handler gets the last write epoch timespamp and
checks against the current epoch to see if session is expired.

Fixes: #151