Added addresses configuration item to the app section.
We can provide a list of addresses where the application will
be listening after the launch.

By default ipv4 and ipv6 loopbacks will be used. This is what
happens when we don't provide an address to the tornado http_server
listen mehtod.

It is better to keep those values on the Firenado configuration
side as we can provide better loggin and use them for other
reasons.

Fixes: #205