The purge call will only handle 500 sesssions per time.
I'm not talking of the number of sessions created but the number
of sessions purged, only 500 per call. If more sessions need to be
purged we'll wait for the next purge wave.

I think that could get tricky on a server a heavy amount of
sessions, maybe should be an good idea add a configuration item
to the purge_limit.

Refs: #221