When purge limit is hit the callback will a hiccup state.
The hiccup state is when there are more sessions to purge than the
purge limit so instead of wait to the next purge callback time out
we go through a hickup timeout until no sessions to purge are
left.

The callback time is 2 minutes(120 secconds) and the callback
hickup time is 10 seconds by default.

With that we delete the purge limit every 10 seconds unblocking the
thread.

With this commit we change session scan_interval config item to
callback_time and callback_hickup is added.

You can set callback_hickup, callback_time and purge_limit session
items in the system and application config files.

Fixes: #221