Introduced process to the app config section.
Now it is possible fork more than one processes via the config.

Setting the app['process']['num_processes'] will fork more than
one processes or infom 0 to fork the number of cpus.

With this commit we add the child indication to the shutdown
messages and fix the wait before shutdown item in the app section.

Also moved the firenado.conf reload to the tornado launcher
constructor as we were missing the wait_before_shutdown value. It
was removed the check for callbacks and timeouts still in the
ioloop as they are not present in the instance during the shutdown
process. Now we just wait for wait_before_shutdown value in
seconds.

Fixes: #264