Configuration load is being handled by cartola.config.
That could break anything using methods to load yaml files from
firenado.config. The framework is handling everything without
problems.

With this commit we import tornado.web in the session module to
prevent the error at the sequence of imports in tornaodweb module.

The issue was session being dependent of the tornado.web import
made in the tornadoweb. Now this is fixed.

Fixes: #355