Used cartola.config instead of importlib.
In the data module we're not even reusing the method from
firenado.config and repeating the same code to resolve classes
from the configuration. Now it is using a cartola.config method.

We renamed all config references at the previous code to conf to
avoid conflict with the module from cartola.

Fixes: #355