Removed type reference from super when calling write_error.
Letting the method resolution order do the job and find
tornado.web.RequestHandler when TornadoHandler is instantiated.

The reason why we're getting this conflict is because
ComponentHandler is being used as a Mixin assuming it's child
will have a write_error method. Adding a type to the super call
would only confure the method resolution order.

The problem here is that a concrete class descending from
ComponentHandler must have the write_error method, in this case
it must inherit from tornado.web.RequestHandler. Maybe this
should be a better solution.

Fixes: #366