Handled error on a non async job.
The job will keep running without breaking the scheduler and job
execution and the full stack trace will be logged with a message
asking to handle the exception.

We don't garantee the job and/or scheduler execution at this point
but we give provide a good point to start to check the root cause
of the exception being thrown during the job execution.

Refs: #349