feat(testing): add process and tornado launcher test cases
The TornadoAsyncTestCase is based on a ProcessLauncher that will
be managed by the test case. With this test case we want to use
async tests to hit the application.

The TornadoAsyncHTTPTestCase is based on a TornadoLauncher that
will provide an application to tornado.testing.AsyncHTTPTestCase
and we can test using the fetch method.

Fixes: #444
Refs: #445